Initial Start
I Copied another Project and startet Integrating the first set of
commands.
Currently the Device is not sending any Status back to ST. I guess this
is what I´m going to set up next. And the Tile-Layout is acting weird.
However, currently supported is turning the device on and off without
receiving a proved feedback of the device